+++ src/test/user.ts
@@ -3,7 +3,7 @@ import request from 'supertest';
 import { should } from 'chai';
 import { app } from '../app';
 import { decode } from 'jsonwebtoken';
-import { DatabaseHelper } from './helper/dbHelper';
+import { DatabaseHelper } from './helper/databaseHelper';
 // TODO: Move to 'expect' instead of 'should'
 should();
